[WIP] DavDocumentsProviderImpl
