+++ app/src/main/kotlin/at/bitfire/davdroid/webdav/DavDocumentsProvider.kt
@@ -10,7 +10,6 @@ import android.content.Context
 import android.content.Intent
 import android.content.res.AssetFileDescriptor
 import android.database.Cursor
-import android.database.MatrixCursor
 import android.graphics.Bitmap
 import android.graphics.BitmapFactory
 import android.graphics.Point
@@ -21,7 +20,6 @@ import android.os.CancellationSignal
 import android.os.ParcelFileDescriptor
 import android.os.storage.StorageManager
 import android.provider.DocumentsContract.Document
-import android.provider.DocumentsContract.Root
 import android.provider.DocumentsContract.buildChildDocumentsUri
 import android.provider.DocumentsContract.buildRootsUri
 import android.provider.DocumentsProvider
@@ -101,6 +99,8 @@ class DavDocumentsProvider(
         fun randomAccessCallbackWrapperFactory(): RandomAccessCallbackWrapper.Factory
         fun streamingFileDescriptorFactory(): StreamingFileDescriptor.Factory
         fun thumbnailCache(): ThumbnailCache
+
+        fun impl(): DavDocumentsProviderImpl
     }
 
     companion object {
@@ -129,7 +129,7 @@ class DavDocumentsProvider(
 
     private val ourContext by lazy { context!! }        // requireContext() requires API level 30
     private val authority by lazy { ourContext.getString(R.string.webdav_authority) }
-    private val entryPoint by lazy { EntryPointAccessors.fromApplication<DavDocumentsProviderEntryPoint>(ourContext) }
+    private lateinit var entryPoint: DavDocumentsProviderEntryPoint
 
     private val logger by lazy { entryPoint.logger() }
 
@@ -153,7 +153,10 @@ class DavDocumentsProvider(
     private val cookieStore by lazy { mutableMapOf<Long, CookieJar>() }
     private val actor by lazy { entryPoint.davDocumentsActorFactory().create(cookieStore, credentialsStore) }
 
-    override fun onCreate() = true
+    override fun onCreate(): Boolean {
+        entryPoint = EntryPointAccessors.fromApplication<DavDocumentsProviderEntryPoint>(context!!)
+        return true
+    }
 
     override fun shutdown() {
         documentProviderScope.cancel()
@@ -162,74 +165,11 @@ class DavDocumentsProvider(
 
     /*** query ***/
 
-    override fun queryRoots(projection: Array<out String>?): Cursor {
-        logger.fine("WebDAV queryRoots")
-        val roots = MatrixCursor(projection ?: arrayOf(
-            Root.COLUMN_ROOT_ID,
-            Root.COLUMN_ICON,
-            Root.COLUMN_TITLE,
-            Root.COLUMN_FLAGS,
-            Root.COLUMN_DOCUMENT_ID,
-            Root.COLUMN_SUMMARY
-        ))
+    override fun queryRoots(projection: Array<out String>?) =
+        entryPoint.impl().queryRoots(projection)
 
-        runBlocking {
-            for (mount in mountDao.getAll()) {
-                val rootDocument = documentDao.getOrCreateRoot(mount)
-                logger.info("Root ID: $rootDocument")
-
-                roots.newRow().apply {
-                    add(Root.COLUMN_ROOT_ID, mount.id)
-                    add(Root.COLUMN_ICON, R.mipmap.ic_launcher)
-                    add(Root.COLUMN_TITLE, ourContext.getString(R.string.webdav_provider_root_title))
-                    add(Root.COLUMN_DOCUMENT_ID, rootDocument.id.toString())
-                    add(Root.COLUMN_SUMMARY, mount.name)
-                    add(Root.COLUMN_FLAGS, Root.FLAG_SUPPORTS_CREATE or Root.FLAG_SUPPORTS_IS_CHILD)
-
-                    val quotaAvailable = rootDocument.quotaAvailable
-                    if (quotaAvailable != null)
-                        add(Root.COLUMN_AVAILABLE_BYTES, quotaAvailable)
-
-                    val quotaUsed = rootDocument.quotaUsed
-                    if (quotaAvailable != null && quotaUsed != null)
-                        add(Root.COLUMN_CAPACITY_BYTES, quotaAvailable + quotaUsed)
-                }
-            }
-        }
-
-        return roots
-    }
-
-    override fun queryDocument(documentId: String, projection: Array<out String>?): Cursor {
-        logger.fine("WebDAV queryDocument $documentId ${projection?.joinToString("+")}")
-
-        val doc = documentDao.get(documentId.toLong()) ?: throw FileNotFoundException()
-        val parent = doc.parentId?.let { parentId ->
-            documentDao.get(parentId)
-        }
-
-        return DocumentsCursor(projection ?: arrayOf(
-            Document.COLUMN_DOCUMENT_ID,
-            Document.COLUMN_DISPLAY_NAME,
-            Document.COLUMN_MIME_TYPE,
-            Document.COLUMN_FLAGS,
-            Document.COLUMN_SIZE,
-            Document.COLUMN_LAST_MODIFIED,
-            Document.COLUMN_ICON,
-            Document.COLUMN_SUMMARY
-        )).apply {
-            val bundle = doc.toBundle(parent)
-            logger.fine("queryDocument($documentId) = $bundle")
-
-            // override display names of root documents
-            if (parent == null) {
-                val mount = runBlocking { mountDao.getById(doc.mountId) }
-                bundle.putString(Document.COLUMN_DISPLAY_NAME, mount.name)
-            }
-
-            addRow(bundle)
-        }
-    }
+    override fun queryDocument(documentId: String, projection: Array<out String>?) =
+        entryPoint.impl().queryDocument(documentId, projection)
 
     /**
      * Gets old or new children of given parent.
